+++ src/threads.js
@@ -2564,6 +2564,10 @@ Process.prototype.reportQuotient = function (a, b) {
     return +a / +b;
 };
 
+Process.prototype.reportPower = function (a, b) {
+    return Math.pow(+a, +b);
+};
+
 Process.prototype.reportModulus = function (a, b) {
     var x = +a,
         y = +b;